Move _relabeled_poly from higherorder to composite
